refactor: it to use the original file
