+++ src/index.js
@@ -1,7 +1,7 @@
 var getData = require('./getData');
 var Sequelize = require('sequelize');
 
-// to see the relevant doc pages we crawl, check indexedFiles.js
+// to see the relevant doc pages we crawl, check indexedFiles.json
 
 // db ops
 var sequelize = new Sequelize('database', null, null, {
